requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-bcrypt>=1.0.1
anthropic>=0.40.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9
Pillow>=10.0
